requirements.txt
wurlitzer
torchvision==0.15.2
jupyterlab
matplotlib
